+++ test/helperFunctions.test.js
